Reorg, add id to resource metadata
