+++ config_test.go
@@ -197,70 +197,6 @@ func TestRemoveResourceNotFoundReturnsError(t *testing.T) {
 	require.Len(t, c.Resources, 6)
 }
 
-func TestParseFQDNParsesComponents(t *testing.T) {
-	fqdn, err := ParseFQDN("module.module1.module2.resource.container.mine.attr")
-	require.NoError(t, err)
-
-	require.Equal(t, "module1.module2", fqdn.Module)
-	require.Equal(t, structs.TypeContainer, fqdn.Type)
-	require.Equal(t, "mine", fqdn.Resource)
-	require.Equal(t, "attr", fqdn.Attribute)
-}
-
-func TestParseFQDNReturnsErrorOnMissingType(t *testing.T) {
-	_, err := ParseFQDN("module.module1.module2.resource.mine")
-	require.Error(t, err)
-}
-
-func TestParseFQDNReturnsErrorOnNoModuleOrResource(t *testing.T) {
-	_, err := ParseFQDN("module1.module2")
-	require.Error(t, err)
-}
-
-func TestParseFQDNReturnsModuleWhenNoResource(t *testing.T) {
-	fqdn, err := ParseFQDN("module.module1.module2")
-	require.NoError(t, err)
-
-	require.Equal(t, "module1.module2", fqdn.Module)
-}
-
-func TestParseFQDNReturnsModuleWhenOutput(t *testing.T) {
-	fqdn, err := ParseFQDN("module.module1.module2.output.mine")
-	require.NoError(t, err)
-
-	require.Equal(t, "module1.module2", fqdn.Module)
-	require.Equal(t, types.TypeOutput, fqdn.Type)
-	require.Equal(t, "mine", fqdn.Resource)
-	require.Equal(t, "value", fqdn.Attribute)
-}
-
-func TestFQDNStringWithoutModuleReturnsCorrectly(t *testing.T) {
-	fqdn, err := ParseFQDN("resource.container.mine")
-	require.NoError(t, err)
-
-	fqdnStr := fqdn.String()
-
-	require.Equal(t, "resource.container.mine", fqdnStr)
-}
-
-func TestFQDNStringWithModuleOutputReturnsCorrectly(t *testing.T) {
-	fqdn, err := ParseFQDN("module.module1.module2.output.mine")
-	require.NoError(t, err)
-
-	fqdnStr := fqdn.String()
-
-	require.Equal(t, "module.module1.module2.output.mine", fqdnStr)
-}
-
-func TestFQDNStringWithModuleResourceReturnsCorrectly(t *testing.T) {
-	fqdn, err := ParseFQDN("module.module1.module2.resource.container.mine")
-	require.NoError(t, err)
-
-	fqdnStr := fqdn.String()
-
-	require.Equal(t, "module.module1.module2.resource.container.mine", fqdnStr)
-}
-
 func TestToJSONSerializesJSON(t *testing.T) {
 	c := testSetupConfig(t)
