+++ types/default.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 )
 
-var TypeNotRegisteredError = fmt.Errorf("type not registered")
+var ErrTypeNotRegistered = fmt.Errorf("type not registered")
 
 type RegisteredTypes map[string]Resource
 
@@ -32,5 +32,5 @@ func (r RegisteredTypes) CreateResource(resourceType, resourceName string) (Reso
 		return res, nil
 	}
 
-	return nil, TypeNotRegisteredError
+	return nil, ErrTypeNotRegistered
 }
